rag-demo1/requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.1
tiktoken>=0.7.0
